requirements.txt
pandas==2.1.3
numpy==1.26.2
matplotlib==3.8.2
seaborn==0.13.0
scikit-learn==1.3.2
tensorflow==2.15.0
Flask-MySQL==1.5.2
mysql-connector-python==8.2.0